Test-Case-Generation-Backend/requirements.txt
# Webservice
python-dotenv
pydantic
pydantic-settings
fastapi
uvicorn
httpx

# LLM
ollama

# Database
supabase

# Memory
redis[hiredis]

# Vector Database
chromadb

# Extras
atlassian-python-api
Werkzeug
authlib
itsdangerous
xlsxwriter